Implemented cancel order functionality for users to cancel their orders.

+++ src/main/java/com/lldproject/blinkitecommerce/repositories/InventoryRepository.java
@@ -10,7 +10,7 @@ import com.lldproject.blinkitecommerce.models.Inventory;
 @Repository
 public interface InventoryRepository extends JpaRepository<Inventory, Integer>{
 
-    Optional<Inventory> findByProduct_Id(int requiredProductId);
+    Optional<Inventory> findByProduct_Id(long requiredProductId);
 
 
 }
